# docs/openapi.yaml
openapi: 3.1.0
info:
  title: MissionPulse API
  version: 1.7.0
  description: |
    MissionPulse API for GovCon proposal management.
    All authenticated endpoints require a valid Supabase JWT session cookie.
    Admin endpoints additionally require admin role RBAC permissions.
  contact:
    name: Mission Meets Tech, LLC
    email: support@missionmeetstech.com
servers:
  - url: https://missionpulse.vercel.app
    description: Production
  - url: http://localhost:3000
    description: Local development
components:
  securitySchemes:
    supabaseAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: Supabase JWT token (passed via session cookie or Authorization header)
    cronSecret:
      type: http
      scheme: bearer
      description: CRON_SECRET bearer token for scheduled job authentication
    stripeSignature:
      type: apiKey
      in: header
      name: stripe-signature
      description: Stripe webhook signature for payload verification
  schemas:
    Error:
      type: object
      properties:
        error:
          type: string
          description: Human-readable error message
      required:
        - error
...
    CronDailyResponse:
      type: object
      properties:
        success:
          type: boolean
        expired_pilots:
          type: integer
          description: Number of pilot subscriptions expired
        engagement_updated:
          type: integer
          description: Number of pilot engagement scores recalculated
        ran_at:
          type: string
          format: date-time
      required:
        - success
        - expired_pilots
        - engagement_updated
        - ran_at
...
security:
  - supabaseAuth: []
paths:
...
  /api/cron/daily:
    get:
      summary: Daily cron job
      description: |
        Processes pilot expiry and engagement scoring. Authenticates via CRON_SECRET bearer token.
        Not for user consumption — called by external scheduler (Vercel Cron, GitHub Actions).
      operationId: runDailyCron
      tags: [Cron]
      security:
        - cronSecret: []
      responses:
        '200':
          description: Cron job completed successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CronDailyResponse'
        '401':
          description: Invalid or missing CRON_SECRET
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '500':
          description: Processing error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
...